format all the test files

+++ src/test/org/epics/archiverappliance/PVCaPut.java
@@ -19,75 +19,75 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
 public class PVCaPut {
-	private static Logger logger = LogManager.getLogger(PVCaPut.class.getName());
+    private static Logger logger = LogManager.getLogger(PVCaPut.class.getName());
 
-          /**
-   * JCA context.
-   */
-  private Context context = null;
-  private JCALibrary jca =null;
-         /**
+    /**
+     * JCA context.
+     */
+    private Context context = null;
+
+    private JCALibrary jca = null;
+    /**
      * Initialize JCA context.
      * @throws CAException      throws on any failure.
-         * @throws ConfigurationException 
-         * @throws IOException 
-         * @throws SAXException 
+     * @throws ConfigurationException
+     * @throws IOException
+     * @throws SAXException
      */
     private void initialize() throws CAException, SAXException, IOException, ConfigurationException, ConfigException {
 
         ConfigServiceForTests configService = new ConfigServiceForTests(-1);
-                // Get the JCALibrary instance.
-         if(jca==null)
-                 jca = JCALibrary.getInstance();
-           
-                        ByteArrayInputStream bis = JCAConfigGen.generateJCAConfig(configService);
-                DefaultConfigurationBuilder configBuilder = new DefaultConfigurationBuilder();
-                Configuration configuration;
-                
-                        configuration = configBuilder.build(bis);
+        // Get the JCALibrary instance.
+        if (jca == null) jca = JCALibrary.getInstance();
 
-                        // Create a context with default configuration values.
-                        if(context==null)
-                                context = jca.createContext(configuration);
+        ByteArrayInputStream bis = JCAConfigGen.generateJCAConfig(configService);
+        DefaultConfigurationBuilder configBuilder = new DefaultConfigurationBuilder();
+        Configuration configuration;
+
+        configuration = configBuilder.build(bis);
+
+        // Create a context with default configuration values.
+        if (context == null) context = jca.createContext(configuration);
     }
-    
-    
+
     /**
      * Destroy JCA context.
      */
     private void destroy() {
-        
+
         try {
 
             // Destroy the context, check if never initialized.
-            if (context != null)
-                context.destroy();
-            
+            if (context != null) context.destroy();
+
         } catch (Throwable t) {
             logger.error("Exception destroying context", t);
         }
     }
-   
-    
-    public void caPut(String pvName,double value) throws CAException, IllegalStateException, TimeoutException, SAXException, IOException, ConfigurationException, ConfigException {
-         initialize();
-          Channel channel = context.createChannel(pvName);
-          context.pendIO(3.0);
-          channel.put(value);
-          context.pendIO(3.0);
-          channel.destroy();      
-          destroy();
+
+    public void caPut(String pvName, double value)
+            throws CAException, IllegalStateException, TimeoutException, SAXException, IOException,
+                    ConfigurationException, ConfigException {
+        initialize();
+        Channel channel = context.createChannel(pvName);
+        context.pendIO(3.0);
+        channel.put(value);
+        context.pendIO(3.0);
+        channel.destroy();
+        destroy();
     }
 
-    public void caPut(String pvName,String value) throws CAException, IllegalStateException, TimeoutException, SAXException, IOException, ConfigurationException, ConfigException {
+    public void caPut(String pvName, String value)
+            throws CAException, IllegalStateException, TimeoutException, SAXException, IOException,
+                    ConfigurationException, ConfigException {
         initialize();
-         Channel channel = context.createChannel(pvName);
-         context.pendIO(3.0);
-         channel.put(value);
-         context.pendIO(3.0);
-         channel.destroy();      
-         destroy();
-   }
+        Channel channel = context.createChannel(pvName);
+        context.pendIO(3.0);
+        channel.put(value);
+        context.pendIO(3.0);
+        channel.destroy();
+        destroy();
+    }
 
     /**
      * caput values into PV
@@ -96,33 +96,31 @@ public class PVCaPut {
      * @param timebetweenupdates - Sleep for this many mills between updates.
      * @throws Exception
      */
-    public void caPutValues(String pvName,double[] values, int timebetweenupdates) throws Exception  {
-    	initialize();
-    	Channel channel = context.createChannel(pvName);
-    	context.pendIO(3.0);
-    	for(double value : values) { 
-        	channel.put(value);
-        	context.pendIO(3.0);
-        	Thread.sleep(timebetweenupdates);
-    	}
-    	channel.destroy();      
-    	destroy();
-   }
-
-        /**
-         * @param args  
-         */
-        public static void main(String[] args) throws Exception {
-                try {
-                        try {
-                                new PVCaPut().caPut("test:enable0", 1);
-                        } catch (SAXException | IOException | ConfigurationException e) {
-                    		logger.error(e);
-                        }
-                } catch (IllegalStateException | CAException | TimeoutException e) {
-            		logger.error(e);
-                }
-
+    public void caPutValues(String pvName, double[] values, int timebetweenupdates) throws Exception {
+        initialize();
+        Channel channel = context.createChannel(pvName);
+        context.pendIO(3.0);
+        for (double value : values) {
+            channel.put(value);
+            context.pendIO(3.0);
+            Thread.sleep(timebetweenupdates);
         }
+        channel.destroy();
+        destroy();
+    }
 
+    /**
+     * @param args
+     */
+    public static void main(String[] args) throws Exception {
+        try {
+            try {
+                new PVCaPut().caPut("test:enable0", 1);
+            } catch (SAXException | IOException | ConfigurationException e) {
+                logger.error(e);
+            }
+        } catch (IllegalStateException | CAException | TimeoutException e) {
+            logger.error(e);
+        }
+    }
 }
